backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
